--- src/components/home/projectsCarousel.js
@@ -1,135 +1,0 @@
-import { useState } from 'react';
-import { Swiper, SwiperSlide } from 'swiper/react';
-import { Autoplay, Navigation, Pagination, Thumbs, A11y } from 'swiper';
-import Link from 'next/link';
-
-import 'swiper/css';
-import 'swiper/css/navigation';
-import 'swiper/css/pagination';
-import 'swiper/css/autoplay';
-import 'swiper/css/thumbs';
-
-export default function ProjectsCarousel() {
-  const thumbnails = [
-    `https://ucarecdn.com/2a8dc700-311d-4bb2-b0d3-8022970754d8/`,
-    `https://ucarecdn.com/6b133fb6-b68b-413c-89a5-54a31a65c66c/`,
-    `https://ucarecdn.com/3ba27e90-3ab4-418f-aa07-05c4ba2c2f55/`,
-    `https://ucarecdn.com/945e1e40-b0e1-40c5-a59d-bc89296cddeb/`,
-    `https://ucarecdn.com/af296ebd-726d-4234-b9ac-449ff7dac3b9/`,
-  ];
-
-  const projectsList = [
-    {
-      name: `SODA for SPARC`,
-      imageUrl: `https://ucarecdn.com/2a8dc700-311d-4bb2-b0d3-8022970754d8/`,
-      description: `SODA (Software to Organize Data Automatically) for SPARC is a desktop software intended to facilitate the data organization and submission process for SPARC investigators according to the FAIR SPARC data standards.`,
-      page: `/sodaforsparc`,
-    },
-    {
-      name: `KnowMore`,
-      imageUrl: `https://ucarecdn.com/6b133fb6-b68b-413c-89a5-54a31a65c66c/`,
-      description: `KnowMore is an automated knowledge discovery tool that allows users of the portal to visualize, in just a few clicks, potential similarities, differences, and connections between multiple SPARC datasets of their choice.`,
-      page: `/knowmore`,
-    },
-    {
-      name: `SPARClink`,
-      imageUrl: `https://ucarecdn.com/3ba27e90-3ab4-418f-aa07-05c4ba2c2f55/`,
-      description: `SPARClink is a system that queries publications using open source tools and platforms and create an interactable visualization that showcases the impact that SPARC and their FAIR data practices have in advancing the field of bioelectronic medicine.`,
-      page: `/sparclink`,
-    },
-    {
-      name: `AQUA`,
-      imageUrl: `https://ucarecdn.com/945e1e40-b0e1-40c5-a59d-bc89296cddeb/`,
-      description: `AQUA (Advanced QUery Architecture for the SPARC Portal) improves the SPARC Portal by making the search engine smarter at understanding user search keywords, enhancing the result display, and providing users with better result filtering and sorting options.`,
-      page: `/aqua`,
-    },
-    {
-      name: `SODA for COVID-19 Research`,
-      imageUrl: `https://i.ibb.co/92qvsnC/imageonline-co-textimage-1.png`,
-      description: `SODA (Sofware to Organize Data Automatically) for COVID-19 Research is a cross-platform desktop software that allows researchers to easily organize and share their COVID-19 related research data according to applicable FAIR guidelines.`,
-      page: `/sodaforcovid`,
-    },
-  ];
-
-  const [SwiperThumbs, setSwiperThumbs] = useState(null);
-
-  return (
-    <>
-      <section>
-        <div className="mx-auto h-auto w-full max-w-screen-lg px-4 pt-1 pb-3">
-          <div className="mb-4 flex flex-col items-center justify-center">
-            <h2 className="my-2 text-center text-4xl font-extrabold tracking-tight sm:text-4xl">
-              Current Projects
-            </h2>
-            <p className="w-full max-w-2xl text-center font-asap text-xl">
-              These are the projects we are working on at the moment:
-            </p>
-          </div>
-
-          <Swiper
-            modules={[Navigation, Pagination, Autoplay, Thumbs, A11y]}
-            slidesPerView={1}
-            navigation
-            setWrapperSize={true}
-            grabCursor={true}
-            loop={true}
-            autoplay={{
-              delay: 6000,
-              pauseOnMouseEnter: true,
-              disableOnInteraction: false,
-            }}
-            mousewheel={{ releaseOnEdges: true }}
-            thumbs={{ swiper: SwiperThumbs }}
-            className="h-full"
-          >
-            {projectsList.map((project, index) => (
-              <SwiperSlide
-                key={index}
-                className="my-auto flex h-full items-center justify-center py-10"
-              >
-                <section className="mx-10 my-auto flex h-full w-4/5 flex-row items-center justify-center rounded-lg px-5 py-10 text-gray-600 shadow-xl">
-                  <div className="container mx-auto flex h-full flex-col items-center justify-center">
-                    <div className="my-5 flex flex-row items-center justify-center sm:py-0 lg:max-w-lg">
-                      <img
-                        className="rounded object-cover object-center sm:h-full sm:w-60 sm:pt-20 md:h-auto md:w-72 md:pt-0 lg:h-auto lg:w-80"
-                        alt="SODA for SPARC logo"
-                        src={project.imageUrl}
-                      />
-                    </div>
-                    <div className="my-4 flex h-full flex-col items-center text-center sm:mb-16 md:mb-0">
-                      <p className="w-full text-left font-asap text-base text-black sm:text-center md:text-base lg:text-lg">
-                        {project.description}
-                      </p>
-
-                      <div className="flex w-full justify-center py-4">
-                        <Link href={project.page} passHref>
-                          <button className="inline-flex rounded border-0 bg-black px-6 py-2 text-lg text-white ring-2 ring-transparent ring-offset-2 transition hover:ring-pink-600 focus:outline-none focus:ring-pink-600 sm:ml-4 md:text-base lg:text-lg">
-                            Learn more about {project.name}
-                          </button>
-                        </Link>
-                      </div>
-                    </div>
-                  </div>
-                </section>
-              </SwiperSlide>
-            ))}
-          </Swiper>
-          <Swiper
-            slidesPerView={5}
-            watchSlidesProgress={true}
-            className="thumbs-swiper hidden sm:block"
-            onSwiper={(swiper) => setSwiperThumbs(swiper)}
-          >
-            {thumbnails.map((thumbnail, index) => (
-              <SwiperSlide virtualIndex={index} key={index}>
-                <div className="thumbnail">
-                  <img src={thumbnail} alt="" />
-                </div>
-              </SwiperSlide>
-            ))}
-          </Swiper>
-        </div>
-      </section>
-    </>
-  );
-}
